Split query condition enums out to tiledb_common crate
