+++ tiledb/common/src/lib.rs
@@ -8,6 +8,7 @@ pub mod datatype;
 pub mod filter;
 pub mod key;
 pub mod metadata;
+pub mod query;
 pub mod range;
 pub mod vfs;
